Add files via upload
In this updated script, the user is prompted to select an action: either installing applications or searching for applications.

If the user selects the "install applications" option, the script prompts the user to enter the names of the applications they want to install, just like in the previous version. The script then confirms the list of applications to be installed and proceeds with the installation if the user confirms.

If the user selects the "search for applications" option, the script prompts the user to enter a search query. The script then searches the Arch Linux repository for packages matching the query and displays the search results to the user.